Update : atom에 named export, 컨벤션 적용, 비사용 컴포넌트 제거
