--- src/atoms/Overlay.tsx
@@ -1,11 +1,0 @@
-import styled from '@emotion/styled';
-
-const Overlay = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-`;
-
-export default Overlay;
