+++ src/atoms/index.ts
@@ -1,13 +1,15 @@
-export { default as TypoAtom } from './TypoAtom';
-export { default as BtnAtom } from './BtnAtom';
-export { default as SwitchAtom } from './SwitchAtom';
-export { default as CardAtom } from './CardAtom';
-export { default as TagAtom, type ITagSpanProps } from './TagAtom';
-export { default as ProgressButtonAtom } from './ProgressButtonAtom';
-export { default as IconAtom } from './IconAtom';
-export { default as Overlay } from './Overlay';
-export { default as InputAtom } from './InputAtom';
-export { default as PopperAtom } from './PopperAtom';
-export { default as TomatoInput } from './TomatoInput';
-export { default as TodoProgressBarAtom } from './TodoProgressBarAtom';
+export { BtnAtom } from './BtnAtom';
+export {
+  type CardAnimationPlayerAnimationType,
+  type ICardAnimationPlayerProps,
+} from './CardAnimationPlayerAtom';
+export { CardAtom } from './CardAtom';
+export { IconAtom } from './IconAtom';
+export { InputAtom } from './InputAtom';
 export { ListAtom } from './ListAtom';
+export { PopperAtom } from './PopperAtom';
+export { ProgressButtonAtom } from './ProgressButtonAtom';
+export { TagAtom, type ITagSpanProps } from './TagAtom';
+export { TodoProgressBarAtom } from './TodoProgressBarAtom';
+export { TomatoInputAtom } from './TomatoInputAtom';
+export { TypoAtom } from './TypoAtom';
